Move StreamEvents from core to otelextension.akka.stream package.
