+++ otel-extension/src/main/java/io/scalac/mesmer/otelextension/akka/MesmerAkkaHelpers.java
@@ -39,8 +39,6 @@ public class MesmerAkkaHelpers {
         "io.scalac.mesmer.core.event.Service$",
         "io.scalac.mesmer.core.event.Service$$anon$1",
         "io.scalac.mesmer.core.event.Service",
-        "io.scalac.mesmer.core.event.StreamEvent$StreamInterpreterStats",
-        "io.scalac.mesmer.core.event.StreamEvent",
         "io.scalac.mesmer.core.util.ClassicActorSystemOps$",
         "io.scalac.mesmer.core.util.ClassicActorSystemOps$ActorSystemOps$",
         "io.scalac.mesmer.core.util.Interval$",
@@ -48,7 +46,6 @@ public class MesmerAkkaHelpers {
         "io.scalac.mesmer.core.util.MutableTypedMap$",
         "io.scalac.mesmer.core.util.MutableTypedMap",
         "io.scalac.mesmer.core.util.Retry$",
-        "io.scalac.mesmer.core.util.stream$",
         "io.scalac.mesmer.core.util.TypedActorSystemOps$",
         "io.scalac.mesmer.core.config.ConfigurationUtils$",
         "io.scalac.mesmer.core.config.ConfigurationUtils$ConfigOps$",
