+++ otel-extension/src/main/java/io/scalac/mesmer/otelextension/akka/MesmerAkkaStreamInstrumentationModule.java
@@ -58,6 +58,12 @@ public class MesmerAkkaStreamInstrumentationModule extends InstrumentationModule
             "akka.stream.GraphLogicOtelOps$GraphLogicEnh$",
             "akka.ConnectionOtelOps",
             "akka.ConnectionOtelOps$",
+            "io.scalac.mesmer.otelextension.instrumentations.akka.stream.StreamService$",
+            "io.scalac.mesmer.otelextension.instrumentations.akka.stream.StreamEvent",
+            "io.scalac.mesmer.otelextension.instrumentations.akka.stream.StreamEvent$",
+            "io.scalac.mesmer.otelextension.instrumentations.akka.stream.StreamEvent$StreamInterpreterStats",
+            "io.scalac.mesmer.otelextension.instrumentations.akka.stream.StreamEvent$LastStreamStats",
+            "io.scalac.mesmer.otelextension.instrumentations.akka.stream.stream$",
             "io.scalac.mesmer.otelextension.instrumentations.akka.stream.impl.ActorGraphInterpreterOtelDecorator",
             "io.scalac.mesmer.otelextension.instrumentations.akka.stream.impl.ActorGraphInterpreterOtelDecorator$",
             "io.scalac.mesmer.otelextension.instrumentations.akka.stream.impl.ActorGraphInterpreterOtelDecorator$$anonfun$addCollectionReceive$1",
