+++ otel-extension/src/main/scala/io/scalac/mesmer/otelextension/instrumentations/akka/stream/impl/ActorGraphInterpreterOtelDecorator.scala
@@ -10,15 +10,15 @@ import akka.stream.GraphLogicOtelOps._
 
 import io.scalac.mesmer.core.akka.model.PushMetrics
 import io.scalac.mesmer.core.event.EventBus
-import io.scalac.mesmer.core.event.StreamEvent.LastStreamStats
-import io.scalac.mesmer.core.event.StreamEvent.StreamInterpreterStats
 import io.scalac.mesmer.core.invoke.Lookup
 import io.scalac.mesmer.core.model.ShellInfo
 import io.scalac.mesmer.core.model.Tag.SubStreamName
 import io.scalac.mesmer.core.model.stream.ConnectionStats
 import io.scalac.mesmer.core.model.stream.StageInfo
-import io.scalac.mesmer.core.util.stream.subStreamNameFromActorRef
 import io.scalac.mesmer.otelextension.instrumentations.akka.stream.AkkaStreamConfig
+import io.scalac.mesmer.otelextension.instrumentations.akka.stream.StreamEvent._
+import io.scalac.mesmer.otelextension.instrumentations.akka.stream.StreamService.streamService
+import io.scalac.mesmer.otelextension.instrumentations.akka.stream.stream.subStreamNameFromActorRef
 
 object ActorGraphInterpreterOtelDecorator extends Lookup {
